Crate iroh_gossip

iroh-gossip

This crate implements the iroh-gossip protocol. It is based on epidemic broadcast trees to disseminate messages among a swarm of peers interested in a topic. The implementation is based on the papers HyParView and PlumTree.

The crate is made up from two modules: The proto module is the protocol implementation, as a state machine without any IO. The net module connects the protocol to the networking stack from iroh-net.

The net module is optional behind the net feature flag (enabled by default).

Getting Started

The iroh-gossip protocol was designed to be used in conjunction with iroh. Iroh is a networking library for making direct connections, these connections are how gossip messages are sent.

Iroh provides a Router that takes an Endpoint and any protocols needed for the application. Similar to a router in webserver library, it runs a loop accepting incoming connections and routes them to the specific protocol handler, based on ALPN.

Here is a basic example of how to set up iroh-gossip with iroh:

use iroh::{protocol::Router, Endpoint};
use iroh_gossip::{net::Gossip, ALPN};
use n0_snafu::ResultExt;

#[tokio::main]
async fn main() -> n0_snafu::Result<()> {
    // create an iroh endpoint that includes the standard discovery mechanisms
    // we've built at number0
    let endpoint = Endpoint::builder().discovery_n0().bind().await?;

    // build gossip protocol
    let gossip = Gossip::builder().spawn(endpoint.clone());

    // setup router
    let router = Router::builder(endpoint.clone())
        .accept(ALPN, gossip.clone())
        .spawn();
    // do fun stuff with the gossip protocol
    router.shutdown().await.e()?;
    Ok(())
}

License

This project is licensed under either of

Apache License, Version 2.0, (LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0)

MIT license (LICENSE-MIT or http://opensource.org/licenses/MIT)

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in this project by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions. Broadcast messages to peers subscribed to a topic

The crate is designed to be used from the iroh crate, which provides a high level interface, but can also be used standalone.

Modules

api

Public API for using iroh-gossip

metrics

Metrics for iroh-gossip

net

Networking for the iroh-gossip protocol

proto

Implementation of the iroh-gossip protocol, as an IO-less state machine

Constants

ALPN

ALPN protocol name