Module iroh_gossip::dispatcher

A higher level wrapper for the gossip engine that manages multiple gossip subscriptions and updates.

Structs

GossipDispatcher

A gossip engine that manages gossip subscriptions and updates.

Message

A gossip message

State 🔒

The mutable state of the gossip engine.

SubscribeOptions

Join a gossip topic

Enums

Command

Send a gossip message

Event

Update from a subscribed gossip topic

GossipEvent

Gossip event An event to be emitted to the application for a particular topic.

TopicState 🔒

Functions

spawn_owned 🔒

tokio::spawn but returns an AbortingJoinHandle that owns the task.

Type Aliases

CommandStream 🔒

Type alias for a stream of gossip updates, so we don’t have to repeat all the bounds.

EventSink 🔒

Type alias for a sink of gossip events.