iroh_gossip::net

Struct Gossip

pub struct Gossip { /* private fields */ }

Publish and subscribe on gossiping topics.

Each topic is a separate broadcast tree with separate memberships.

A topic has to be joined before you can publish or subscribe on the topic. To join the swarm for a topic, you have to know the PublicKey of at least one peer that also joined the topic.

Messages published on the swarm will be delivered to all peers that joined the swarm for that topic. You will also be relaying (gossiping) messages published by other peers.

With the default settings, the protocol will maintain up to 5 peer connections per topic.

Even though the Gossip is created from a Endpoint, it does not accept connections itself. You should run an accept loop on the Endpoint yourself, check the ALPN protocol of incoming connections, and if the ALPN protocol equals GOSSIP_ALPN, forward the connection to the gossip actor through Self::handle_connection.

The gossip actor will, however, initiate new connections to other peers by itself.

Implementations

impl Gossip

pub fn builder() -> Builder

Creates a default Builder, with the endpoint set.

pub fn max_message_size(&self) -> usize

Get the maximum message size configured for this gossip actor.

pub async fn handle_connection(&self, conn: Connection) -> Result<(), Error>

Handle an incoming Connection.

Make sure to check the ALPN protocol yourself before passing the connection.

pub async fn shutdown(&self) -> Result<(), Error>

Shutdown the gossip instance.

This leaves all topics, sending Disconnect messages to peers, and then stops the gossip actor loop and drops all state and connections.

pub fn metrics(&self) -> &Arc<Metrics>

Returns the metrics tracked for this gossip instance.