iroh_gossip::net

Type Alias ProtoCommand

pub type ProtoCommand = Command<PublicKey>;

Commands for the gossip protocol

Aliased Type

enum ProtoCommand {
    Join(Vec<PublicKey>),
    Broadcast(Bytes, Scope),
    Quit,
}

Variants

Join(Vec<PublicKey>)

Join this topic and connect to peers.

If the list of peers is empty, will prepare the state and accept incoming join requests, but only become operational after the first join request by another peer.

Broadcast(Bytes, Scope)

Broadcast a message for this topic.

Quit

Leave this topic and drop all state.