Module iroh_gossip::proto::hyparview

Implementation of the HyParView membership protocol

The implementation is based on this paper by Joao Leitao, Jose Pereira, Luıs Rodrigues and the example implementation by Bartosz Sypytkowski

Structs

Config

Configuration for the swarm membership layer

Disconnect

Message sent when leaving the swarm or closing down to inform peers about us being gone.

ForwardJoin

A message informing other peers that a new peer joined the swarm for this topic.

Neighbor

A neighbor message is sent after adding a peer to our active view to inform them that we are now neighbors.

Shuffle

Shuffle messages are sent occasionally to shuffle our passive view with peers from other peer’s active and passive views.

ShuffleReply

Once a shuffle messages reaches a Ttl of 0, a peer replies with a ShuffleReply.

State

The state of the HyParView protocol

Stats

Ttl

The time-to-live for this message.

Enums

Event

Event emitted by the State to the application.

InEvent

Input event for HyParView

Message

Messages that we can send and receive from peers within the topic.

OutEvent

Output event for HyParView

Priority

The priority of a Join message

RemovalReason 🔒

Timer

Kinds of timers HyParView needs to schedule.

Type Aliases

Alive

Respond