Struct iroh_gossip::proto::hyparview::State

pub struct State<PI, RG = ThreadRng> {
    me: PI,
    me_data: Option<PeerData>,
    pub(crate) active_view: IndexSet<PI>,
    pub(crate) passive_view: IndexSet<PI>,
    config: Config,
    shuffle_scheduled: bool,
    rng: RG,
    stats: Stats,
    pending_neighbor_requests: HashSet<PI>,
    peer_data: HashMap<PI, PeerData>,
}

The state of the HyParView protocol

Fields
me: PI

Our peer identity

me_data: Option<PeerData>

Our opaque user data to transmit to peers on join messages

active_view: IndexSet<PI>

The active view, i.e. peers we are connected to

passive_view: IndexSet<PI>

The passive view, i.e. peers we know about but are not connected to at the moment

config: Config

Protocol configuration (cannot change at runtime)

shuffle_scheduled: bool

Whether a shuffle timer is currently scheduled

rng: RG

Random number generator

stats: Stats

Statistics

pending_neighbor_requests: HashSet<PI>

The set of neighbor requests we sent out but did not yet receive a reply for

peer_data: HashMap<PI, PeerData>

The opaque user peer data we received for other peers

Implementations

impl<PI, RG> State<PI, RG>
where PI: PeerIdentity, RG: Rng,

pub fn new(me: PI, me_data: Option<PeerData>, config: Config, rng: RG) -> Self

pub fn handle(&mut self, event: InEvent<PI>, now: Instant, io: &mut impl IO<PI>)

fn handle_message( &mut self, from: PI, message: Message<PI>, now: Instant, io: &mut impl IO<PI> )

fn handle_join(&mut self, peer: PI, io: &mut impl IO<PI>)

fn handle_disconnect(&mut self, peer: PI, io: &mut impl IO<PI>)

fn handle_quit(&mut self, io: &mut impl IO<PI>)

fn on_join( &mut self, peer: PI, data: Option<PeerData>, now: Instant, io: &mut impl IO<PI> )

fn on_forward_join( &mut self, sender: PI, message: ForwardJoin<PI>, now: Instant, io: &mut impl IO<PI> )

fn on_neighbor( &mut self, from: PI, details: Neighbor, now: Instant, io: &mut impl IO<PI> )

fn peer_info(&self, id: &PI) -> PeerInfo<PI>

Get the peer PeerInfo for a peer.

fn insert_peer_info(&mut self, peer_info: PeerInfo<PI>, io: &mut impl IO<PI>)

fn on_shuffle(&mut self, from: PI, shuffle: Shuffle<PI>, io: &mut impl IO<PI>)

Handle a Message::Shuffle

A node q that receives a Shuffle request will first decrease its time to live. If the time to live of the message is greater than zero and the number of nodes in q’s active view is greater than 1, the node will select a random node from its active view, different from the one he received this shuffle message from, and simply forwards the Shuffle request. Otherwise, node q accepts the Shuffle request and send back (p.8)

fn on_shuffle_reply(&mut self, message: ShuffleReply<PI>, io: &mut impl IO<PI>)

fn on_disconnect(&mut self, peer: PI, details: Disconnect, io: &mut impl IO<PI>)

fn handle_shuffle_timer(&mut self, io: &mut impl IO<PI>)

fn passive_is_full(&self) -> bool

fn active_is_full(&self) -> bool

fn add_passive( &mut self, peer: PI, data: Option<PeerData>, io: &mut impl IO<PI> )

Add a peer to the passive view.

If the passive view is full, it will first remove a random peer and then insert the new peer. If a peer is currently in the active view it will not be added.

fn remove_active( &mut self, peer: &PI, respond: bool, io: &mut impl IO<PI> ) -> Option<PI>

Remove a peer from the active view.

If respond is true, a Disconnect message will be sent to the peer.

fn refill_active_from_passive( &mut self, skip_peers: &[&PI], io: &mut impl IO<PI> )

fn handle_pending_neighbor_timer(&mut self, peer: PI, io: &mut impl IO<PI>)

fn remove_active_by_index( &mut self, peer_index: usize, respond: bool, reason: RemovalReason, io: &mut impl IO<PI> ) -> Option<PI>

fn free_random_slot_in_active_view(&mut self, io: &mut impl IO<PI>)

Remove a random peer from the active view.

fn add_active( &mut self, peer: PI, data: Option<PeerData>, priority: Priority, _now: Instant, io: &mut impl IO<PI> ) -> bool

Add a peer to the active view.

If the active view is currently full, a random peer will be removed first. Sends a Neighbor message to the peer. If high_priority is true, the peer may not deny the Neighbor request.

fn add_active_unchecked( &mut self, peer: PI, priority: Priority, io: &mut impl IO<PI> )