Module iroh_gossip::proto::plumtree

Implementation of the Plumtree epidemic broadcast tree protocol

The implementation is based on this paper by Joao Leitao, Jose Pereira, Luıs Rodrigues and the example implementation by Bartosz Sypytkowski

Structs

Config

Configuration for the gossip broadcast layer.

Gossip

Payload messages transmitted by the protocol.

GossipEvent

Graft

Control message to signal a peer that they have been moved to the eager set, and to ask the peer to do the same with this node.

IHave

Control message to inform peers we have a message without transmitting the whole payload.

MessageId

A message identifier, which is the message content’s blake3 hash.

Round

Number of delivery hops a message has taken.

State

State of the plumtree.

Stats

Stats about this topic’s plumtree.

Enums

DeliveryScope

The scope to deliver the message to.

Event

Event emitted by the State to the application.

InEvent

Events Plumtree is informed of from the peer sampling service and IO layer.

Message

Messages that we can send and receive from peers within the topic.

OutEvent

Events Plumtree emits.

Scope

The broadcast scope of a gossip message.

Timer

Kinds of timers Plumtree needs to schedule.