Struct iroh_gossip::proto::plumtree::State

pub struct State<PI> {
    me: PI,
    config: Config,
    pub(crate) eager_push_peers: HashSet<PI>,
    pub(crate) lazy_push_peers: HashSet<PI>,
    lazy_push_queue: HashMap<PI, Vec<IHave>>,
    missing_messages: HashMap<MessageId, VecDeque<(PI, Round)>>,
    received_messages: TimeBoundCache<MessageId, ()>,
    cache: TimeBoundCache<MessageId, Gossip>,
    graft_timer_scheduled: HashSet<MessageId>,
    dispatch_timer_scheduled: bool,
    init: bool,
    pub(crate) stats: Stats,
}

State of the plumtree.

Fields
me: PI

Our address.

config: Config

Configuration for this plumtree.

eager_push_peers: HashSet<PI>

Set of peers used for payload exchange.

lazy_push_peers: HashSet<PI>

Set of peers used for control message exchange.

lazy_push_queue: HashMap<PI, Vec<IHave>>missing_messages: HashMap<MessageId, VecDeque<(PI, Round)>>

Messages for which a MessageId has been seen via a Message::IHave but we have not yet received the full payload. For each, we store the peers that have claimed to have this message.

received_messages: TimeBoundCache<MessageId, ()>

Messages for which the full payload has been seen.

cache: TimeBoundCache<MessageId, Gossip>

Payloads of received messages.

graft_timer_scheduled: HashSet<MessageId>

Message ids for which a Timer::SendGraft has been scheduled.

dispatch_timer_scheduled: bool

Whether a Timer::DispatchLazyPush has been scheduled.

init: bool

Set to false after the first message is received. Used for initial timer scheduling.

stats: Stats

Stats of this plumtree.

Implementations

impl<PI: PeerIdentity> State<PI>

pub fn new(me: PI, config: Config) -> Self

Initialize the State of a plumtree.

pub fn handle(&mut self, event: InEvent<PI>, now: Instant, io: &mut impl IO<PI>)

Handle an InEvent.

pub fn stats(&self) -> &Stats

Get access to the Stats of the plumtree.

fn handle_message( &mut self, sender: PI, message: Message, now: Instant, io: &mut impl IO<PI> )

Handle receiving a Message.

fn on_dispatch_timer(&mut self, io: &mut impl IO<PI>)

Dispatches messages from lazy queue over to lazy peers.

fn broadcast( &mut self, content: Bytes, scope: Scope, now: Instant, io: &mut impl IO<PI> )

Send a gossip message.

Will be pushed in full to eager peers. Pushing the message id to the lazy peers is delayed by a timer.

fn on_gossip( &mut self, sender: PI, message: Gossip, now: Instant, io: &mut impl IO<PI> )

Handle receiving a Message::Gossip.

fn optimize_tree( &mut self, gossip_sender: &PI, message: &Gossip, previous_ihaves: VecDeque<(PI, Round)>, io: &mut impl IO<PI> )

Optimize the tree by pruning the sender of a Message::Gossip if we previously received a Message::IHave for the same message with a much lower number of delivery hops from the original broadcaster of the message.

See Config::optimization_threshold.

fn on_prune(&mut self, sender: PI)

Handle receiving a Message::Prune.

fn on_ihave(&mut self, sender: PI, ihaves: Vec<IHave>, io: &mut impl IO<PI>)

Handle receiving a Message::IHave.

When a node receives a IHAVE message, it simply marks the corresponding message as missing It then starts a timer, with a predefined timeout value, and waits for the missing message to be received via eager push before the timer expires. The timeout value is a protocol parameter that should be configured considering the diameter of the overlay and a target maximum recovery latency, defined by the application requirements. This is a parameter that should be statically configured at deployment time. (p8)

fn on_send_graft_timer(&mut self, id: MessageId, io: &mut impl IO<PI>)

A scheduled Timer::SendGraft has reached it’s deadline.

fn on_graft(&mut self, sender: PI, details: Graft, io: &mut impl IO<PI>)

Handle receiving a Message::Graft.

fn on_neighbor_up(&mut self, peer: PI)

Handle a InEvent::NeighborUp when a peer joins the topic.

fn on_neighbor_down(&mut self, peer: PI)

Handle a InEvent::NeighborDown when a peer leaves the topic.

When a neighbor is detected to leave the overlay, it is simple removed from the membership. Furthermore, the record of IHAVE messages sent from failed members is deleted from the missing history. (p9)

fn on_evict_cache_timer(&mut self, now: Instant, io: &mut impl IO<PI>)

fn add_eager(&mut self, peer: PI)

Moves peer into eager set.

fn add_lazy(&mut self, peer: PI)

Moves peer into lazy set.

fn eager_push(&mut self, gossip: Gossip, sender: &PI, io: &mut impl IO<PI>)

Immediately sends message to eager peers.

fn lazy_push(&mut self, gossip: Gossip, sender: &PI, io: &mut impl IO<PI>)

Queue lazy message announcements into the queue that will be sent out as batched Message::IHave messages once the Timer::DispatchLazyPush timer is triggered.