Struct iroh_gossip::proto::util::IndexSet

pub(crate) struct IndexSet<T> {
    inner: IndexSet<T>,
}

A hash set where the iteration order of the values is independent of their hash values.

This is wrapper around [indexmap::IndexSet] which couple of utility methods to randomly select elements from the set.

Fields
inner: IndexSet<T>
Implementations

impl<T: Hash + Eq + PartialEq> IndexSet<T>

pub fn new() -> Self

pub fn insert(&mut self, value: T) -> bool

pub fn remove_random<R: Rng + ?Sized>(&mut self, rng: &mut R) -> Option<T>

Remove a random element from the set.

pub fn pick_random<R: Rng + ?Sized>(&self, rng: &mut R) -> Option<&T>

Pick a random element from the set.

pub fn pick_random_without<R: Rng + ?Sized>( &self, without: &[&T], rng: &mut R ) -> Option<&T>

Pick a random element from the set, but not any of the elements in without.

pub fn pick_random_index<R: Rng + ?Sized>(&self, rng: &mut R) -> Option<usize>

Pick a random index for an element in the set.

pub fn remove(&mut self, value: &T) -> Option<T>

Remove an element from the set.

NOTE: the value is removed by swapping it with the last element of the set and popping it off. This modifies the order of element by moving the last element

pub fn remove_index(&mut self, index: usize) -> Option<T>

Remove an element from the set by its index.

NOTE: the value is removed by swapping it with the last element of the set and popping it off. This modifies the order of element by moving the last element

pub fn iter_without<'a>(&'a self, value: &'a T) -> impl Iterator<Item = &'a T>

Create an iterator over the set in the order of insertion, while skipping the element in without.

impl<T> IndexSet<T>
where T: Hash + Eq + Clone,

pub fn shuffled<R: Rng + ?Sized>(&self, rng: &mut R) -> Vec<T>

Create a vector of all elements in the set in random order.

pub fn shuffled_and_capped<R: Rng + ?Sized>( &self, len: usize, rng: &mut R ) -> Vec<T>

Create a vector of all elements in the set in random order, and shorten to the first len elements after shuffling.

pub fn shuffled_without<R: Rng + ?Sized>( &self, without: &[&T], rng: &mut R ) -> Vec<T>

Create a vector of the elements in the set in random order while omitting the elements in without.

pub fn shuffled_without_and_capped<R: Rng + ?Sized>( &self, without: &[&T], len: usize, rng: &mut R ) -> Vec<T>

Create a vector of the elements in the set in random order while omitting the elements in without, and shorten to the first len elements.