iroh_gossip::proto

Module topic

This module contains the implementation of the gossiping protocol for an individual topic

Structs

Config

Protocol configuration

State

The topic state maintains the swarm membership and broadcast tree for a particular topic.

Stats

Statistics for the protocol state of a topic

Enums

Command

A command to the protocol state for a particular topic.

Event

An event to be emitted to the application for a particular topic.

InEvent

Input event to the topic state handler.

Message

A protocol message for a particular topic

OutEvent

An output event from the state handler.

Timer

A timer to be registered for a particular topic.

Traits

IO

A trait for a concrete type to push OutEvents to.